feat(culture): restructuring folder on culture feature
