+++ lib/features/culture/data/models/culture_model/common_model.dart
@@ -1,4 +1,4 @@
-import 'package:boole_apps/features/culture/domain/entities/common_entity.dart';
+import 'package:boole_apps/features/culture/domain/entities/culture_entities/common_entity.dart';
 
 class CommonModel extends Common {
   CommonModel({required super.phrase, required super.translation});
