+++ lib/features/culture/data/models/culture_model/dress_code_model.dart
@@ -1,4 +1,5 @@
-import 'package:boole_apps/features/culture/domain/entities/dress_code_entity.dart';
+
+import 'package:boole_apps/features/culture/domain/entities/culture_entities/dress_code_entity.dart';
 
 class DressCodeModel extends DressCode {
   DressCodeModel({required super.men, required super.women});
