+++ lib/features/culture/data/models/culture_model/social_etiquette_model.dart
@@ -1,4 +1,5 @@
-import 'package:boole_apps/features/culture/domain/entities/social_etiquette_entity.dart';
+
+import 'package:boole_apps/features/culture/domain/entities/culture_entities/social_etiquette_entity.dart';
 
 class SocialEtiquetteModel extends SocialEtiquette {
   SocialEtiquetteModel({required super.family, required super.greetings});
