+++ lib/features/culture/data/models/culture_model/traditional_arts_model.dart
@@ -1,4 +1,5 @@
-import 'package:boole_apps/features/culture/domain/entities/traditional_arts_entity.dart';
+
+import 'package:boole_apps/features/culture/domain/entities/culture_entities/traditional_arts_entity.dart';
 
 class TraditionalArtsModel extends TraditionalArts {
   TraditionalArtsModel({required super.dance, required super.crafts});
