+++ lib/features/culture/domain/entities/culture_entities/cultural_highlights_entity.dart
