+++ lib/features/culture/domain/entities/culture_entities/social_etiquette_entity.dart
