+++ lib/features/culture/domain/repositories/culture_repository.dart
@@ -1,4 +1,5 @@
-import 'package:boole_apps/features/culture/domain/entities/culture_entity.dart';
+
+import 'package:boole_apps/features/culture/domain/entities/culture_entities/culture_entity.dart';
 
 abstract class CultureRepository {
   Future<List<Culture>> getCulture();
